Changed DeleteMe's location
